fix(trusty-audit): split the selection file out of run.rs, repoint two Test citations (#5823)
Three CI gates, all mechanical:

- `run.rs` reached 555 SLOC, over the 500 production cap. The selection file is
  the seam that costs nothing to cut: its format, its two write obligations and
  the reader that enforces them are one subject, and its producers (#5497,
  `crate::clone`) live outside this module. Moved to `run/selection.rs` and
  re-exported, so every caller's path is unchanged.
- Two `Test:` pointers in tga's relay cited a test name that was renamed before
  it landed.
- `trusty-progress` 0.3.1 is published and this PR adds a public module to it,
  so its version moves to 0.3.2 — additive, which is the patch position under
  Cargo's 0.x rule.
