+++ crates/trusty-progress/Cargo.toml
@@ -4,7 +4,7 @@ description = "Shared rustup/cargo-style progress + status display for the trust
 homepage = "https://github.com/bobmatnyc/trusty-tools"
 keywords = ["progress", "spinner", "cli", "terminal", "status"]
 categories = ["command-line-interface", "development-tools"]
-version = "0.3.1"
+version = "0.3.2"
 edition = "2021"
 rust-version.workspace = true
 license.workspace = true
